Move InputMode to helpers to avoid overlapping var and module names
